breaks tasks into modules
